Add support for clone depth
Clone depth is specified by the "depth" setting
in the manifest. If not specified, the default value can
be set by the grit clone option "--depth".
Similarly, the default value for "single-branch" can now
also be set by the grit clone option "--single-branch".

Also, added support for "--no-post-run" to skip running
the "run-after-clone" commands in the manifest.